# src/swagger/swagger.json
{
    "swagger": "2.0",
    "host": "localhost:3000",
    "basePath": "/",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/movie": {
            "post": {
                "tags": [
                    "movie"
                ],
                "summary": "Add a new movie",
                "description": "",
                "operationId": "addMovie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "movie created",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Movie"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Movie"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "actor not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Movie": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "releaseYear": {
                    "type": "string"
                },
                "plot": {
                    "type": "string"
                },
                "poster": {
                    "type": "string"
                },
                "actor_id": {
                    "type": "string"
                },
                "producer_id": {
                    "type": "string"
                }
            }
        }
    }
}